Move copy_results method out of big_ep_helpers
