+++ src/dlstbx/util/big_ep_helpers.py
@@ -2,13 +2,9 @@ from __future__ import annotations
 
 import json
 import os
-import shutil
-import subprocess
-from copy import deepcopy
 from itertools import tee
 
 import libtbx.load_env
-import py
 from cctbx.eltbx import henke, sasaki
 from cctbx.sgtbx import space_group, space_group_symbols
 from cctbx.uctbx import unit_cell
@@ -431,41 +427,6 @@ def ispyb_write_model_json(working_directory, mdl_dict, logger):
         os.path.join(working_directory, "big_ep_model_ispyb.json"), "w"
     ) as json_file:
         json_file.write(json_data)
-
-
-def copy_results(working_directory, results_directory, skip_copy, logger):
-    def ignore_func(directory, files):
-        ignore_list = deepcopy(skip_copy)
-        pth = py.path.local(directory)
-        for f in files:
-            fp = pth.join(f)
-            if not fp.check():
-                ignore_list.append(f)
-                continue
-            if os.path.islink(fp):
-                dest = os.readlink(fp)
-                if not os.path.isfile(dest):
-                    ignore_list.append(f)
-        return ignore_list
-
-    shutil.copytree(
-        working_directory,
-        results_directory,
-        symlinks=False,
-        ignore_dangling_symlinks=True,
-        ignore=ignore_func,
-    )
-    src_pth_esc = r"\/".join(os.path.dirname(working_directory).split(os.sep))
-    dest_pth_esc = r"\/".join(os.path.dirname(results_directory).split(os.sep))
-    sed_command = (
-        r"find %s -type f -exec grep -Iq . {} \; -and -exec sed -i 's/%s/%s/g' {} +"
-        % (results_directory, src_pth_esc, dest_pth_esc)
-    )
-    logger.info(f"Running sed command: {sed_command}")
-    try:
-        subprocess.call([sed_command], shell=True)
-    except Exception:
-        logger.warning("Failed to run sed command to update paths", exc_info=True)
 
 
 def send_results_to_ispyb(results_directory, log_files, record_result):
